Ci/workflow telemetry (#86)
Adding workflow telemetry for the local integration test. This shows a realistic view in terms of memory consumption of the local env which comes out to 3-4GB RAM, which is not bad! I think we have like a single core on a Github action, so that should be fine. The disk is a pessimistic estimate as this job builds the actual images, but probably we can do better than this: around 9GB usage.