Highlight from tensor network (#128)
We can highlight outer legs now for tensornetworks from the calculated parity check matrix.
I also reworked the "caching" for weight enums and parity check matrix to be instead states stored in the tensornetwork zustand store. This is a better "object oriented" design, where we retrieve the data based on the tensornetwork signature. This is also leading towards a better design for solving the refresh problem #41.

Fixes #3.