Mypy completeness + WEP keys are now simple Pauli lists (#137)
All the classes are now mypy type checked. Added pylint for further linting and catching errors, and in preparation for enforcing public methods/classes/modules to have docstrings.

Also reworked the key type in the tensor enumerator to be Pauli representation based not symplectic, thus giving a ~20-30% speedup to the WEP calculations, as now the slicing / comparison of keys are much cheaper.

Fixes #33.