Merged contraction paths (#214)
Improves the contraction algorithm by merging contraction of multiple legs into a single step. 
We also extract a `Tracable` interface and the `Contraction` class, abstracting out the contraction behavior that was shared in a naive way previously between the `conjoin_nodes` and the `stabilizer_weight_enumerator` features. Custom cost function implementation is adjusted and simplified accordingly.